requirements.txt
aiogram
python-dotenv
deep-translator
langdetect
aiohttp
